gmail-to-sheet-automation/requirements.txt
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
